web/src/requirements.txt
Flask==2.0.0
mysql-connector-python
python-dotenv